requirements.txt
streamlit>=1.36
neo4j>=5.20
openai>=1.40
python-dotenv>=1.0
